'List files' options; cleanup; etc.
Some cleanup; added '-l' option (list files, don't play them) and '-L' option
(list files and then play them) and a bit of CL arg error processing.  The
file list is piped to a pager via popen so that the user won't have to
watch a long list scroll off the screen.

Also, a small bug fix: Enclose file names in double quotes instead of
single quotes so that a single quote in a file name won't cause an error -
although a double quote in a file name will still be a problem.