Append-to/rebuild database logic, force, etc.
Added logic to, if the cache/database file exists and a command-line arg
matches no files, run 'locate' on that arg and append the result to the
database file (in case new audio files have been created after the
database file was created).  Included a bit of refactoring: moved the
logic to load the file-name arrays from the database file into a new
routine: load_in_memory_db.

Added -f option to force complete rebuilding of the cache/database file.
Changed 'require' to require_relative to allow ruby to find the needed
modules when playmusic is run from another directory.