Changes for rubocop and a couple other improvements
configuration.rb: Implemented "lazy" initialization of the DBFile so that a
new DB file will never be created if the user gives the -h (help) option.

playmusic: Moved call to configuration.db to after 'if opt.cl_error' so
that the lazy initialization of the DBFile (described above) will not occur
if the user specified the -h option (or some invalid option).  When the
pager is used via IO.popen, changed the error handling logic to ignore the
Errno::EPIPE error, since a broken pipe is usually not a problem from the
user's point of view (i.e., they hit 'q' to stop the pager and don't care
that there was more input to view/page).

general changes: modifications to please rubocop (the largest being
creating some utility routines to get rid of "Method has too many lines"
warning); some small, mostly cosmetic, improvements.